backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-CORS==4.0.0
groq==0.4.2
